Kassandra Staking and Gov first commit
There are still things to be wrapped up, but the main concepts of how it will work are already there.

The main contract is `Staking.sol`, it's inspired by Synthetix staking rewards, Wault Finance multi-pool staking storage and Sovryn voting power system.
The governance-related functions are in the `StakingGov.sol`, it's based on the Compound token and Sovryn voting power system.
Nonetheless, `StakingStorage.sol` stores the data needed for the staking contract to work.
Others contracts could be further removed and imported from open libraries as OpenZepellin.